Move arrayUtils into libstrings.js
Prior to #23297 it needed to be in its own file, but that is no longer
that case.
